recruiting-tool/requirements.txt
boto3>=1.34.0
python-jobspy>=1.1.0
